requirements.txt
pandas==2.1.4
openpyxl==3.1.2
openai==1.6.1
pyyaml==6.0.1
jinja2==3.1.2
requests==2.31.0
python-dotenv==1.0.0

